move server to zkvm-host bin
