+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [ "aggregation-client",
-    "crates/*", "host-server",
+    "crates/*",
     "native-host",
     "validity-client",
     "zkvm-client",
