--- host-server/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "host-server"
-version = "0.1.0"
-license.workspace = true
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-axum = "0.7.4"
-
-# workspace
-tokio = { workspace = true }
-serde = { workspace = true }
-
-# kona
-kona-host = { workspace = true }
-
-# local
-host-utils.workspace = true
-
-# sp1
-sp1-sdk = { workspace = true }
-dotenv.workspace = true
-anyhow.workspace = true
