+++ zkvm-host/Cargo.toml
@@ -19,6 +19,10 @@ path = "bin/multi.rs"
 name = "agg"
 path = "bin/agg.rs"
 
+[[bin]]
+name = "server"
+path = "bin/server.rs"
+
 [dependencies]
 
 # workspace
@@ -39,6 +43,10 @@ dotenv.workspace = true
 num-format.workspace = true
 revm.workspace = true
 client-utils.workspace = true
+serde = { workspace = true }
+
+# server
+axum = "0.7.4"
 
 [build-dependencies]
 sp1-helper = { workspace = true }
